+++ CHANGELOG.md
@@ -2,5 +2,8 @@
 All notable changes to this project will be documented in this file. This change log follows the conventions of [keepachangelog.com](http://keepachangelog.com/).
 
 ## [Unreleased]
+### Added
+- `if->` and `when->`, and paired thread-last versions to complement `cond->`
+- `when-pred`, `if-pred` and `cond-pred` functions for thread pipelines branching on the current value of the expression
 
 [Unreleased]: https://github.com/IGJoshua/more-threads/compare/a08cb7c68faa4b6d3bfb64c5129fcacc19a4efb4...HEAD